+++ python/paddle/apy/matmul_pass/matmul/cutlass_patch/all_tuning_configs.h
@@ -511,4 +511,10 @@ struct GemmTuningConfigs<float, SwizzleFactor, Batched, 12> {
   static constexpr int kId = 12;
 };
 
+struct DefaultConfig {
+  static constexpr int kConfigId = 0;
+  static constexpr int kSwizzleFactor = 1;
+  static constexpr bool kBatched = false;
+};
+
 }  // namespace ap
